@@ -3,30 +3,34 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.24",
-    "@types/lodash": "^4.14.173",
-    "@types/node": "^12.20.25",
-    "@types/react": "^17.0.21",
-    "@types/react-dom": "^17.0.9",
-    "axios": "^0.22.0",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.11.3",
+    "@mui/material": "^5.6.3",
+    "@mui/styles": "^5.6.2",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/user-event": "^14.1.1",
+    "@types/jest": "^27.4.1",
+    "@types/lodash": "^4.14.182",
+    "@types/node": "^17.0.27",
+    "@types/react": "^18.0.7",
+    "@types/react-dom": "^18.0.0",
+    "axios": "^0.27.1",
     "classnames": "^2.3.1",
     "final-form": "^4.20.2",
     "js-cookie": "^3.0.1",
     "lodash": "^4.17.21",
-    "ra-language-czech": "^2.0.0",
-    "react": "^17.0.2",
-    "react-admin": "^3.18.3",
-    "react-dom": "^17.0.2",
-    "react-final-form": "^6.5.3",
-    "react-router": "^5.2.1",
-    "react-scripts": "^4.0.3",
-    "typescript": "^4.5.2",
-    "web-vitals": "^1.1.2"
+    "ra-language-czech": "^3.19.8",
+    "react": "^18.0.0",
+    "react-admin": "^4.0.1",
+    "react-dom": "^18.0.0",
+    "react-final-form": "^6.5.9",
+    "react-router": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "typescript": "^4.6.3",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -53,6 +57,6 @@
     ]
   },
   "devDependencies": {
-    "@types/react-router": "^5.1.16"
+    "@types/react-router": "^5.1.18"
   }
 }
