@@ -1,12 +1,134 @@
-import React from 'react'
+import React, { ReactNode, createElement } from 'react';
+import lodashGet from 'lodash/get';
+import clsx from 'clsx';
+import { useResourceDefinitions, useGetResourceLabel, useCreatePath,} from 'ra-core';
+import { DRAWER_WIDTH, CLOSED_DRAWER_WIDTH, useSidebarState, DashboardMenuItem, MenuItemLink } from 'react-admin';
+import PropTypes from 'prop-types';
+import { styled } from '@mui/material/styles';
+
+import AddIcon from '@mui/icons-material/Add'
+import DefaultIcon from '@mui/icons-material/ViewList';
+
+const Menu = (props: MenuProps) => {
+
+    const resources = useResourceDefinitions();
+    const getResourceLabel = useGetResourceLabel();
+    const createPath = useCreatePath();
+
+    const {
+        hasDashboard,
+        dense,
+        children = (
+            <>
+                {hasDashboard && <DashboardMenuItem dense={dense} />}
+                {Object.keys(resources)
+                    .filter(name => resources[name].hasList)
+                    .map(name => (
+                        <MenuItemLink
+                            key={name}
+                            to={createPath({
+                                resource: name,
+                                type: 'list',
+                            })}
+                            state={{ _scrollToTop: true }}
+                            primaryText={getResourceLabel(name, 2)}
+                            leftIcon={
+                                resources[name].icon ? (
+                                    createElement(resources[name].icon)
+                                ) : (
+                                    <DefaultIcon />
+                                )
+                            }
+                            dense={dense}
+                        />
+                    ))}
+                <MenuItemLink
+                    key="/PossessionSettings/create"
+                    to="/PossessionSettings/create"
+                    state={{ _scrollToTop: true }}
+                    primaryText="Přidat"
+                    leftIcon={<AddIcon />}
+                    dense={dense}
+                />
+            </>
+        ),
+        className,
+        ...rest
+    } = props;
+
+    const [open] = useSidebarState();
+
+    return (
+        <Root
+            className={clsx(
+                {
+                    [MenuClasses.open]: open,
+                    [MenuClasses.closed]: !open,
+                },
+                className
+            )}
+            {...rest}
+        >
+            {children}
+        </Root>
+    );
+};
+
+export interface MenuProps {
+    children?: ReactNode;
+    className?: string;
+    dense?: boolean;
+    hasDashboard?: boolean;
+}
+
+Menu.propTypes = {
+    className: PropTypes.string,
+    dense: PropTypes.bool,
+    hasDashboard: PropTypes.bool,
+};
+
+const PREFIX = 'RaMenu';
+
+export const MenuClasses = {
+    open: `${PREFIX}-open`,
+    closed: `${PREFIX}-closed`,
+};
+
+const Root = styled('div', {
+    name: PREFIX,
+    overridesResolver: (props, styles) => styles.root,
+})(({ theme }) => ({
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'flex-start',
+    marginTop: '0.5em',
+    marginBottom: '1em',
+    [theme.breakpoints.only('xs')]: {
+        marginTop: 0,
+    },
+    transition: theme.transitions.create('width', {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.leavingScreen,
+    }),
+
+    [`&.${MenuClasses.open}`]: {
+        width: lodashGet(theme, 'sidebar.width', DRAWER_WIDTH),
+    },
+
+    [`&.${MenuClasses.closed}`]: {
+        width: lodashGet(theme, 'sidebar.closedWidth', CLOSED_DRAWER_WIDTH),
+    },
+}));
+
+export default Menu
+/*import React from 'react'
 import { FC, ReactNode } from 'react'
 import PropTypes from 'prop-types'
-import { shallowEqual, useSelector } from 'react-redux'
-import { useMediaQuery, Theme } from '@material-ui/core'
-import { makeStyles } from '@material-ui/core/styles'
+import { useMediaQuery, Theme } from '@mui/material'
+import { makeStyles } from '@mui/styles'
 import DefaultIcon from '@material-ui/icons/ViewList'
-import AddIcon from '@material-ui/icons/Add'
-import { DashboardMenuItem, MenuItemLink, useGetResourceLabel, getResources, ReduxState } from 'react-admin'
+import AddIcon from '@mui/icons-material/Add'
+import { DashboardMenuItem, MenuItemLink, useGetResourceLabel, getResources } from 'react-admin'
 import classnames from 'classnames'
 import { get as lodashGet } from 'lodash'
 
@@ -131,4 +253,4 @@ Menu.defaultProps = {
     onMenuClick: () => null,
 };
 
-export default Menu
+export default Menu*/
