@@ -1,5 +1,5 @@
 import React from 'react'
-import UsersIcon from '@material-ui/icons/Settings'
+import UsersIcon from '@mui/icons-material/Settings'
 
 const LazyCreate = React.lazy(() => import('./Create'))
 const LazyEdit = React.lazy(() => import('./Edit'))
@@ -7,7 +7,7 @@ const LazyShow = React.lazy(() => import('./Show'))
 const LazyList = React.lazy(() => import('./List'))
 
 const articleSettingsProps = {
-    name: 'ArticleSettings',
+    name: 'PossessionSettings',
     icon: UsersIcon,
     list: LazyList,
     show: LazyShow,
