@@ -1,5 +1,7 @@
 import React, { useEffect, useState } from 'react'
 import { Admin, Resource as RaResource } from 'react-admin'
+import { QueryClient } from 'react-query'
+import axios from 'axios'
 
 import userProps from './ra-models/User'
 import permissionProps from './ra-models/Permission'
@@ -7,27 +9,33 @@ import roleProps from './ra-models/Role'
 import typeSettingsProps from './ra-models/ArticleSettings'
 
 import Dashboard from './ra-customs/Dashboard'
+import Resource from './ra-customs/Resource_v4'
 import Layout from './ra-customs/Layout'
-import customRoutes from './ra-customs/customRoutes'
-import Resource from './ra-customs/Resource'
 
 import dataProvider from './providers/dataProvider'
 import authProvider from './providers/authProvider'
 import i18nProvider from './providers/I18nProvider'
 
-import { getCreateComponent, getEditComponent, getListComponent, getShowComponent } from './utils/utils'
-
 import './App.css'
-import axios from 'axios'
+
+import { getCreateComponent, getEditComponent, getListComponent, getShowComponent } from './utils/utils'
 
 const apiUrl = 'http://localhost:5000/api/'
 
 const App: React.FC = () => {
 
+    const queryClient = new QueryClient({
+        defaultOptions: {
+            queries: {
+                staleTime: 1 * 60 * 1000, // 1 minute
+            },
+        },
+    });
+
     const [articleSettings, setArticleSettings] = useState([])
 
     useEffect(() => {
-        axios.get(apiUrl + 'articleSettings', {
+        axios.get(apiUrl + 'possessionSettings', {
             headers: {
                 //"Access-Control-Expose-Headers": "Content-Range"
             },
@@ -35,14 +43,13 @@ const App: React.FC = () => {
         })
             .then(resp => {
                 console.log(resp)
-                // @ts-ignore
                 setArticleSettings(resp.data.data)
             })
     }, [])
 
     const resources = articleSettings.map((obj: any) => {
         return (
-            <RaResource 
+            <RaResource
                 key={obj.name}
                 name={obj.name}
                 list={getListComponent(obj)}
@@ -58,9 +65,9 @@ const App: React.FC = () => {
             dataProvider={dataProvider}
             authProvider={authProvider}
             i18nProvider={i18nProvider}
+            queryClient={queryClient}
             dashboard={Dashboard}
             layout={Layout}
-            customRoutes={customRoutes}
         >
             {resources}
             <Resource {...userProps} />
